requirements_hf.txt
# AI Business Tools - Hugging Face Spaces Requirements
# Optimized for HF Spaces deployment with open source models

# Core
streamlit==1.28.0
python-dotenv==1.0.0

# LLM & Agents
langchain==0.1.0
langchain-core==0.1.0
langchain-community==0.0.1
langchain-ollama==0.0.1

# Hugging Face integration
huggingface-hub==0.19.0
transformers==4.35.0

# Data & Processing
pandas==2.0.0
numpy==1.24.0
pydantic==2.0.0

# Utilities
requests==2.31.0
python-dateutil==2.8.0

# Optional: for exports
reportlab==4.0.0
openpyxl==3.1.0
